+++ src/adapters/edit-auth-info.js
