+++ src/components/sign-in/sign-in.js
@@ -11,10 +11,10 @@ class SignIn extends React.PureComponent {
     this.handleSubmit = this.handleSubmit.bind(this);
   }
 
-  handleSubmit() {
-    const onSubmit = this.props;
+  handleSubmit(evt) {
+    const {onSubmit} = this.props;
 
-    // evt.preventDefault();
+    evt.preventDefault();
 
     onSubmit({
       login: this.loginRef.current.value,
