+++ src/reducer/operation.test.js
@@ -3,7 +3,7 @@ import MockAdapter from "axios-mock-adapter";
 import {createApi} from "../api";
 import {ActionTypes as DataActionTypes} from "./data/data";
 import {ActionTypes as AppActionTypes} from "./app/app";
-import EditOffer from "../editOffer";
+import EditOffer from "../adapters/edit-offer";
 
 const api = createApi(() => {});
